requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# A2A SDK and related
google-genai==0.2.1
anthropic==0.7.7

# MCP support
mcp==0.9.0

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
